requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Werkzeug
WTForms
email-validator
SQLAlchemy
python-dotenv
requests
json
Pillow
psycopg2
psycopg2-binary
sqlite3-to-mysql
pymysql
gunicorn
black